Correct client validation
In light of discussion in slack, the problem here is that the validation
is based on the *negotiated* version.  Validation needs to be based on
the client's *selected* version.

I've also moved text about validating lack of support for the original
version to the incompatible VN section, alongside a reminder to validate
connection IDs.  Security does depend on this, but the client ignores
these bad VN packets.  That provides more effective defense against
attack than what was proposed.